unit_type_add_bill

+++ src/main/java/lk/grocery/pos/tm/OrderItemDetailTM.java
@@ -1,14 +1,12 @@
 package lk.grocery.pos.tm;
 
-import javafx.scene.control.Button;
-
 import java.io.Serializable;
 import java.math.BigDecimal;
 
 public class OrderItemDetailTM implements Serializable {
     private String itemCode;
     private String itemName;
-    private int quantity;
+    private double quantity;
     private BigDecimal unitPrice;
     private String unitType;
     private BigDecimal discount;
@@ -17,7 +15,7 @@ public class OrderItemDetailTM implements Serializable {
     public OrderItemDetailTM() {
     }
 
-    public OrderItemDetailTM(String itemCode, String itemName, int quantity, BigDecimal unitPrice, String unitType, BigDecimal discount, BigDecimal total) {
+    public OrderItemDetailTM(String itemCode, String itemName, double quantity, BigDecimal unitPrice, String unitType, BigDecimal discount, BigDecimal total) {
         this.itemCode = itemCode;
         this.itemName = itemName;
         this.quantity = quantity;
@@ -42,11 +40,11 @@ public class OrderItemDetailTM implements Serializable {
         this.itemName = itemName;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
